+++ flask_restly/_storage.py
@@ -49,7 +49,7 @@ def push_mapping(func, path, serialize, method):
     metadata.get(parent_name)[func.__name__] = data
 
 
-def push_body_types(func, incoming, outgoing):
+def push_metadata(func, new_data):
     parent_name = _get_func_parent_name(func)
     metadata = get_metadata_storage()
 
@@ -60,32 +60,7 @@ def push_body_types(func, incoming, outgoing):
 
     data = metadata.get(parent_name).get(func.__name__, {}).copy()
 
-    data.update(
-        {
-            'incoming': incoming,
-            'outgoing': outgoing,
-        }
-    )
-
-    metadata.get(parent_name)[func.__name__] = data
-
-
-def push_skip_authorization(func):
-    parent_name = _get_func_parent_name(func)
-    metadata = get_metadata_storage()
-
-    if not any(parent_name == key for key in metadata.keys()):
-        metadata.set(parent_name, {
-            func.__name__: {},
-        })
-
-    data = metadata.get(parent_name).get(func.__name__, {}).copy()
-
-    data.update(
-        {
-            'skip_authorization': True,
-        }
-    )
+    data.update(new_data)
 
     metadata.get(parent_name)[func.__name__] = data