+++ flask_restly/decorator/queued.py
@@ -1,0 +1,14 @@
+from flask_restly._storage import push_metadata
+from functools import wraps
+
+
+def queued(func):
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        return func(*args, **kwargs)
+
+    push_metadata(wrapper, {
+        'queued': True,
+    })
+
+    return wrapper